requirements.txt
torch>=1.9.0
transformers>=4.40.0
datasets>=2.19.0
huggingface_hub>=0.23.0
fsspec>=2024.3.0
accelerate
sentencepiece
sacrebleu
bert_score
PyYAML
numpy